# application/services/custom_services/nutrition/src/models/nutritionfood.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const nutritionfoodSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   name: String
})

const nutritionfoodModel = mongoose.model('nutritionfood', nutritionfoodSchema, 'nutritionfood');
export default nutritionfoodModel;